Customizable slot merge strategies.
Previously the only supported merge strategy was to compare for equality
and declare a conflict if any states are unequal. Although that remains
the default, we can now provide a custom merge strategy to allow for
different use cases, like a slot containing a set that we only ever add
to, whose merge strategy is to combine all of the sets together using
the union operation.