Allow slots to be manually finalized.
Sometimes a caller will decide that they are happy with the present state
of a slot and decide to finalize it early. For example, the slot may
be relevant only to one sub-section of the program that the datafork
module can't "see".

This allows the caller to manually finalize a given slot, causing it to
freeze at its current value (that is, the value according to the
currently-active state) at the time of the call. Just as with the previous
implicit finalization, the slot becomes read-only after this and it
no longer depends on its associated root or the current state in order
to deliver its value/positions.